This commit renames all files and directories in the repository to use the snake_case naming convention. It also updates all corresponding references in the code, such as `#include` statements, to reflect these changes.
