+++ src/pi_controller.cpp
@@ -1,4 +1,4 @@
-#include "PIController.h"
+#include "pi_controller.h"
 
 // A utility function for constraining a value within a range
 template<typename T>
